requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.21
PyMySQL==1.1.0
python-dotenv==1.0.0
requests==2.31.0
Jinja2==3.1.2
Werkzeug==2.3.7
pyserial==3.5
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.21
PyMySQL==1.1.0
python-dotenv==1.0.0
requests==2.31.0
Jinja2==3.1.2
Werkzeug==2.3.7
